@@ -1,5 +1,6 @@
+# typed: strict
 # frozen_string_literal: true
 
 module BrickdocSettings
-  VERSION = "0.0.1"
+  VERSION = '0.0.1'
 end
